requirements.txt
astroid==1.6.4
autopep8==1.3.5
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.6
django-filter==2.0.0
gunicorn==19.8.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.1.0
psycopg2-binary==2.7.4
pycodestyle==2.4.0
pylint==1.9.1
python-decouple==3.1
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
